requirements.txt
pytest==8.2.2
requests==2.32.4
pytest-html==4.1.1
configparser==7.0.0
selenium==4.22.0
webdriver-manager==4.0.2
faker==37.5.3
allure-pytest==2.15.0
tenacity==9.1.2
httpx==0.28.1
